handle closing

+++ pkg/stream/producer.go
@@ -302,9 +302,12 @@ func (producer *Producer) FlushUnConfirmedMessages() {
 }
 
 func (producer *Producer) Close() error {
+	if producer.getStatus() == closed {
+		return AlreadyClosed
+	}
 	producer.setStatus(closed)
 	if !producer.options.client.socket.isOpen() {
-		return fmt.Errorf("connection already closed")
+		return fmt.Errorf("tcp connection is closed")
 	}
 
 	err := producer.options.client.deletePublisher(producer.ID)
